refactor(runtimed): extract LaunchKernel handler
Moves the 1.3k-line LaunchKernel arm out of the dispatcher into
`requests/launch_kernel.rs`. Behavior is identical — env-source resolution,
project-file detection, inline-dep promotion, captured-env reuse, pool-env
acquisition, RestartKernel vs spawn path, and runtime-agent handshake all
move verbatim.

Helpers used only by LaunchKernel are lifted to `pub(crate)`:
`resolve_metadata_snapshot`, `captured_env_source_override`,
`captured_env_for_runtime`, `unified_env_on_disk`, `build_launched_config`,
`check_inline_deps`, `get_inline_{uv,conda}_deps`,
`get_inline_uv_prerelease`, `get_inline_conda_channels`,
`extract_pixi_toml_deps`, `try_{uv,conda}_pool_for_inline_deps`,
`acquire_prewarmed_env_with_capture`, `promote_inline_deps_to_project`,
`reset_starting_state`, `publish_kernel_state_presence`. The `CapturedEnv`
and `CapturedEnvRuntime` enums along with `runtime_agent_generation` and
`pending_runtime_agent_connect_tx` on `NotebookRoom` also go pub(crate).

Lock scoping is preserved exactly: every `state_doc`/`doc`/`runtime_agent_*`
guard still drops at the same block boundary it did before, so the
no-tokio-mutex-across-await lint still passes.
